Add logging.

+++ src/main/java/com/creek/whereareyoumodel/domain/Identifiable.java
@@ -5,8 +5,8 @@ package com.creek.whereareyoumodel.domain;
  * @author Andrey Pereverzin
  */
 public interface Identifiable {
-    int getId();
-    void setId(int id);
+    long getId();
+    void setId(long id);
     ContactCompoundId getContactCompoundId();
     void setContactCompoundId(ContactCompoundId contactCompoundId);
 }
